Reorganize windows settings

+++ zzz-windows/debloat.ps1
@@ -7,54 +7,71 @@ $apps_exclude = @(
     # Required to install things...
     'Microsoft.WindowsStore'
 )
-$features_exclude = @(
-)
 $capabilities_exclude = @(
     # Cannot be removed
     'Language.Basic~*',
-    'Language.TextToSpeech~*'
+
+    # Cannot be removed in some versions (should check)
+    'Language.TextToSpeech~*',
+
+    # Stuff I generally want to keep
+    'Microsoft.Windows.MSPaint~*',
+    'Microsoft.Windows.Notepad~*'
+)
+$features_exclude = @(
 )
 
 # Remove all apps
-$apps_prov = Get-AppxProvisionedPackage -Online |
-    where {$i=$_; -not ($apps_exclude | where {$i.DisplayName -like $_})}
-$apps_prov | Format-Table
-$apps_prov | Remove-AppxProvisionedPackage -Online
 $apps = Get-AppxPackage |
     where {$i=$_; -not ($apps_exclude | where {$i.Name -like $_})}
 $apps | Format-Table
-$apps | Remove-AppxPackage -ErrorAction SilentlyContinue
+$prompt = Read-Host -Prompt 'Remove apps? (n)'
+if ($prompt -eq 'y') {
+    $apps | Remove-AppxPackage -ErrorAction SilentlyContinue
+}
+
+# Remove all provisioned apps
+$apps_prov = Get-AppxProvisionedPackage -Online |
+    where {$i=$_; -not ($apps_exclude | where {$i.DisplayName -like $_})}
+$apps_prov | Format-Table
+$prompt = Read-Host -Prompt 'Remove provisioned apps? (n)'
+if ($prompt -eq 'y') {
+    $apps_prov | Remove-AppxProvisionedPackage -Online
+}
+
+# Remove all capabilities
+$capabilities = Get-WindowsCapability -Online | where State -eq 'Installed' |
+    where {$i=$_; -not ($capabilities_exclude | where {$i.Name -like $_})}
+$capabilities | Format-Table
+$prompt = Read-Host -Prompt 'Remove capabilities? (n)'
+if ($prompt -eq 'y') {
+    $capabilities | Remove-WindowsCapability -Online
+}
 
 # Remove all optional features
 $features = Get-WindowsOptionalFeature -Online | where State -eq 'Enabled' |
     where {$i=$_; -not ($features_exclude | where {$i.FeatureName -like $_})}
 $features | Format-Table
-$features | Disable-WindowsOptionalFeature -Online -NoRestart
-
-# Remove all optional capabilities
-$capabilities = Get-WindowsCapability -Online | where State -eq 'Installed' |
-    where {$i=$_; -not ($capabilities_exclude | where {$i.Name -like $_})}
-$capabilities | Format-Table
-$capabilities | Remove-WindowsCapability -Online
-
-# Remove OneDrive
-$uninstall = Get-Package -Name "Microsoft OneDrive" -ProviderName Programs -ErrorAction Ignore | ForEach-Object -Process {$_.Meta.Attributes["UninstallString"]}
-if ($uninstall) {
-    echo "Uninstalling OneDrive"
-    taskkill /f /im OneDrive.exe
-    iex "& $uninstall"
-    pause
-    Stop-Process -Name explorer -Force
-    Remove-Item "$env:LocalAppData\Microsoft\OneDrive" -Force -Recurse
-    Remove-Item "$env:ProgramData\Microsoft OneDrive" -Force -Recurse
+$prompt = Read-Host -Prompt 'Disable optional features? (n)'
+if ($prompt -eq 'y') {
+    $features | Disable-WindowsOptionalFeature -Online -NoRestart
 }
 
-# Remove Edge
-# NOTE: It keeps reinstalling itself
-#.\lib\uninstall_edge.ps1
+# Remove bloatware
+$prompt = Read-Host -Prompt 'Uninstall OneDrive? (n)'
+if ($prompt -eq 'y') {
+    .\lib\uninstall_onedrive.ps1
+}
+$prompt = Read-Host -Prompt 'Uninstall Edge? (n)'
+if ($prompt -eq 'y') {
+    .\lib\uninstall_edge.ps1
+}
 
 # Unpin everything from the start menu
-.\lib\unpin_start.ps1
+$prompt = Read-Host -Prompt 'Unpin all applications? (n)'
+if ($prompt -eq 'y') {
+    .\lib\unpin_start.ps1
+}
 
 # Reboot...
 echo "Time to reboot"
